C_Simple_Calculator

+++ C_Simple_Calculator.cpp
@@ -2,10 +2,12 @@
 using namespace std;
 int main()
 {
-    int a, b;
+    long long int a, b;
     cin >> a >> b;
+    // long long int multiplication = a * b;
     cout << a << " + " << b << " = " << a + b << endl;
     cout << a << " * " << b << " = " << a * b << endl;
+    // cout << a << " * " << b << " = " << multiplication << endl;
     cout << a << " - " << b << " = " << a - b << endl;
     return 0;
 }
